Move frosted glass component to main compose scope
